pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "db_exp"
authors = [{ name = "Henning Krause", email = "henning.krause90@gmail.com" }]
license = { file = "LICENSE" }
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python =">=3.10.6"


dependencies = [
    "azure-identity",
    "azure-keyvault-secrets",
    "asyncpg",
    "fastapi",
    "uvicorn",
    "psycopg2-binary",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "black",
    "pylint"
]
